+++ rusk/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Remove legacy event system 
 
+### Fixed
+
+- Fix support for testnet cluster [#TBD]
+
 ## [1.0.2] - 2025-01-27
 
 ### Added